delete duplicated wasm module
